backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
pytest==8.1.0
pytest-asyncio==0.23.8
